+++ src/runtime/switcher.c
@@ -12,6 +12,7 @@
  * would be to bucketize the futexes and let tasks select the bucket based on
  * the hash of the ids.
  ******************************************************************************/
+#include <inttypes.h>
 #include <stdbool.h>
 #include <stdio.h>
 // clang-format off
@@ -61,7 +62,7 @@ switcher_status_t
 switcher_yield(task_id id, bool (*any_task_filter)(task_id))
 {
     task_id prev, next;
-    logger_debugf("YIELD  task %lu\n", id);
+    logger_debugf("YIELD  task %" PRIu64 "\n", (uint64_t)id);
 
     vmutex_acquire(&_switcher.mutex);
     int bucket = (int)(id % LOTTO_SWITCHER_NBUCKETS);
@@ -111,7 +112,7 @@ switcher_yield(task_id id, bool (*any_task_filter)(task_id))
     vmutex_release(&_switcher.mutex);
 
     _lotto_switcher_resuming();
-    logger_debugf("RESUME task %lu\n", id);
+    logger_debugf("RESUME task %" PRIu64 "\n", (uint64_t)id);
 
     return status;
 }
@@ -122,7 +123,7 @@ void
 switcher_wake(task_id id, nanosec_t slack)
 {
     vmutex_acquire(&_switcher.mutex);
-    logger_debugf("WAKE   task %lu\n", id);
+    logger_debugf("WAKE   task %" PRIu64 "\n", (uint64_t)id);
 
     ASSERT(_switcher.next == NO_TASK);
     ASSERT(id != NO_TASK);
